move method to multishop repo
